resolve merge issues
